Support for a richer overrides format
As the code stands today, the overrides feature only accepts version
as an override. This is extremely limiting as there may be times where
it would be desirable to override other aspects of the software's
configuration. This change allows one to now pass in a yaml or json
file for overriding the properties of the software class.